add tools.py para funções e troca de biblioteca para leitura de dbf
